python/monitoramentos/codigos/requirements.txt
# Dependências do Sistema de Monitoramentos

# Banco de dados PostgreSQL
psycopg2-binary>=2.9.0

# Variáveis de ambiente
python-dotenv>=0.19.0

# Processamento de dados (opcional, para debug)
pandas>=1.3.0

# Logging estruturado (opcional)
structlog>=21.0.0

